demoprac/src/test/java/com/imtiyaaz/tpapppractical/Factories/AccountFactoryTest.java
package com.imtiyaaz.tpapppractical.Factories;

import com.imtiyaaz.tpapppractical.Domain.Account;
import com.imtiyaaz.tpapppractical.Domain.Client;
import com.imtiyaaz.tpapppractical.Factories.AccountFactory;
import org.junit.Before;
import org.junit.Test;

import java.util.HashMap;
import java.util.List;
import java.util.Map;

import static org.junit.Assert.assertEquals;

/**
 * Created by Imtiyaaz on 2017/08/13.
 */
public class AccountFactoryTest {
    Map<String, String> values;
    @Before
    public void setUp() throws Exception {
        values = new HashMap<String, String>();
        List<Client> ClientList;

        values.put("AccountType", "Savings");
        values.put("Date", " ");
        values.put("AccountNumber", "1234567");
    }

    @Test
    public void testGetAccount() throws Exception {
        Account account = AccountFactory.getAccount(values, 200.0);
        System.out.println(account.getAccountNumber());
        assertEquals("23456", account.getAccountNumber());
    }

}